backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20

# Database
sqlalchemy>=2.0.0
alembic>=1.14.0
asyncpg>=0.30.0
psycopg2-binary>=2.9.10

# Validation & config
pydantic[email]>=2.10.0
pydantic-settings>=2.7.0

# HTTP client (proxy forwarding)
httpx>=0.28.0

# Cache & queues
redis>=5.2.0

# Security
cryptography>=44.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# Token counting
tiktoken>=0.8.0

# Logging
structlog>=25.0.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.25.0
